SRB2: MSVC still have a cast for it's broken arm? oh boy
git-svn-id: svn://code.srb2.org/SRB2/trunk@5271 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/sdl/i_sound.c
@@ -521,7 +521,7 @@ static inline void I_SetChannels(void)
 	for (i = -128; i < 128; i++)
 	{
 		const double po = pow((double)(2.0l), (double)((long double)i/1024.0l));
-		steptablemid[i] = po;
+		steptablemid[i] = (INT32)po;
 	}
 
 	// Generates volume lookup tables
